+++ src/MOCK_DB_DATA/library-items/library-items-shared-data.ts
@@ -167,10 +167,18 @@ export const LibraryItemsSharedData: LibraryItemsShared = {
   },
   TimeSelectorWeek: {
     id: 'time-selector-week',
-    title: 'Week selector'
+    title: 'Weekpicker'
+  },
+  TimeSelectorWeekRange: {
+    id: 'time-selector-week-range',
+    title: 'Week range'
   },
   TimeSelectorYear: {
     id: 'time-selector-year',
     title: 'Year selector'
+  },
+  TimeSelectorYearRange: {
+    id: 'time-selector-year-range',
+    title: 'Year range'
   },
 };