+++ src/MOCK_DB_DATA/library-items/time-selectors/_time-selectors.data.ts
@@ -7,15 +7,18 @@ import { TimeSelectorDatepicker } from './time-selector-datepicker';
 import { TimeSelectorDateAndTime } from './time-selector-date-and-time';
 import { TimeSelectorDateRange } from './time-selector-date-range';
 import { TimeSelectorWeek } from './time-selector-week';
+import { TimeSelectorWeekRange } from './time-selector-week-range';
 import { TimeSelectorYear } from './time-selector-year';
+import { TimeSelectorYearRange } from './time-selector-year-range';
 
 const TimeSelectors: LibraryItem[] = [
   ...TimeSelectorDatepicker,
   ...TimeSelectorDateAndTime,
   ...TimeSelectorWeek,
-  ...TimeSelectorDateRange,
+  ...TimeSelectorWeekRange,
   ...TimeSelectorYear,
   ...TimeSelectorDateRange,
+  ...TimeSelectorYearRange
 ];
 
 export const TimeSelectorsData: LibraryItemGroup = {